rafactor mapping os strategy params
