+++ frontend/src/component/feature/FeatureView/FeatureOverview/FeatureOverviewEnvironments/FeatureOverviewEnvironment/EnvironmentAccordionBody/StrategyDraggableItem/StrategyItem/StrategyExecution/StrategyExecution.tsx
@@ -1,31 +1,15 @@
-import {
-    Children,
-    isValidElement,
-    useMemo,
-    type FC,
-    type ReactNode,
-} from 'react';
+import { Children, isValidElement, type FC, type ReactNode } from 'react';
 import { styled } from '@mui/material';
-import type {
-    CreateFeatureStrategySchema,
-    StrategySchema,
-    StrategySchemaParametersItem,
-} from 'openapi';
+import type { CreateFeatureStrategySchema } from 'openapi';
 import type { IFeatureStrategyPayload } from 'interfaces/strategy';
 import { useUiFlag } from 'hooks/useUiFlag';
 import { StrategyExecution as LegacyStrategyExecution } from './LegacyStrategyExecution';
-import { StrategyExecutionItem } from './StrategyExecutionItem/StrategyExecutionItem';
 import { ConstraintItem } from './ConstraintItem/ConstraintItem';
 import { useStrategies } from 'hooks/api/getters/useStrategies/useStrategies';
 import { objectId } from 'utils/objectId';
-import { StrategyChip } from './StrategyChip/StrategyChip';
-import {
-    parseParameterNumber,
-    parseParameterString,
-    parseParameterStrings,
-} from 'utils/parseParameter';
 import { StrategyExecutionSeparator } from './StrategyExecutionSeparator/StrategyExecutionSeparator';
-import { Truncator } from 'component/common/Truncator/Truncator';
+import { useCustomStrategyParameters } from './hooks/useCustomStrategyItems';
+import { useStrategyParameters } from './hooks/useStrategyParameters';
 
 const StyledGrayscale = styled('div', {
     shouldForwardProp: (prop) => prop !== 'enabled',
@@ -76,190 +60,6 @@ const ListItem: FC<{ children: ReactNode }> = ({ children }) => (
     <StyledListItem>{children}</StyledListItem>
 );
 
-const RolloutParameter: FC<{
-    value?: string | number;
-    parameters?: (
-        | IFeatureStrategyPayload
-        | CreateFeatureStrategySchema
-    )['parameters'];
-    hasConstraints?: boolean;
-    displayGroupId?: boolean;
-}> = ({ value, parameters, hasConstraints, displayGroupId }) => {
-    const percentage = parseParameterNumber(value);
-
-    const explainStickiness =
-        typeof parameters?.stickiness === 'string' &&
-        parameters?.stickiness !== 'default';
-    const stickiness = explainStickiness ? (
-        <>
-            with <strong>{parameters.stickiness}</strong>
-        </>
-    ) : (
-        ''
-    );
-
-    return (
-        <StrategyExecutionItem type='Rollout %'>
-            <StrategyChip label={`${percentage}%`} /> of your base {stickiness}
-            <span>
-                {hasConstraints ? 'who match constraints ' : ' '}
-                is included.
-            </span>
-            {/* TODO: displayGroupId */}
-        </StrategyExecutionItem>
-    );
-};
-
-const useStrategyParameters = (
-    strategy: IFeatureStrategyPayload | CreateFeatureStrategySchema,
-    displayGroupId?: boolean,
-) => {
-    const { constraints } = strategy;
-    const { parameters } = strategy;
-    const hasConstraints = Boolean(constraints?.length);
-    const parameterKeys = parameters ? Object.keys(parameters) : [];
-    const mapPredefinedStrategies = (key: string) => {
-        if (key === 'rollout' || key === 'Rollout') {
-            return (
-                <RolloutParameter
-                    key={key}
-                    value={parameters?.[key]}
-                    parameters={parameters}
-                    hasConstraints={hasConstraints}
-                    displayGroupId={displayGroupId}
-                />
-            );
-        }
-
-        if (
-            ['userIds', 'UserIds', 'hostNames', 'HostNames', 'IPs'].includes(
-                key,
-            )
-        ) {
-            return (
-                <StrategyExecutionItem
-                    key={key}
-                    type={key}
-                    values={parseParameterStrings(parameters?.[key])}
-                />
-            );
-        }
-
-        return null;
-    };
-
-    return useMemo(
-        () => parameterKeys.map(mapPredefinedStrategies).filter(Boolean),
-        [parameters, hasConstraints, displayGroupId],
-    );
-};
-
-const useCustomStrategyItems = (
-    strategy: IFeatureStrategyPayload | CreateFeatureStrategySchema,
-    strategies: StrategySchema[],
-) => {
-    const { parameters } = strategy;
-    const definition = useMemo(
-        () =>
-            strategies.find((strategyDefinition) => {
-                return strategyDefinition.name === strategy.name;
-            }),
-        [strategies, strategy.name],
-    );
-    const isCustomStrategy = definition?.editable;
-
-    const mapCustomStrategies = (
-        param: StrategySchemaParametersItem,
-        index: number,
-    ) => {
-        if (!parameters || !param.name) return null;
-        const { type, name } = param;
-        const typeItem = <Truncator title={name}>{name}</Truncator>;
-        const key = `${type}${index}`;
-
-        switch (type) {
-            case 'list': {
-                const values = parseParameterStrings(parameters[name]);
-                if (!values || values.length === 0) {
-                    return null;
-                }
-
-                return (
-                    <StrategyExecutionItem
-                        key={key}
-                        type={typeItem}
-                        values={values}
-                    >
-                        {values.length === 1
-                            ? 'has 1 item:'
-                            : `has ${values.length} items:`}
-                    </StrategyExecutionItem>
-                );
-            }
-
-            case 'percentage': {
-                const value = parseParameterNumber(parameters[name]);
-                return (
-                    <StrategyExecutionItem key={key} type={typeItem}>
-                        is set to <StrategyChip label={`${value}%`} />
-                    </StrategyExecutionItem>
-                );
-            }
-
-            case 'boolean': {
-                const value = parameters[name];
-                return (
-                    <StrategyExecutionItem key={key} type={typeItem}>
-                        is set to <StrategyChip label={value} />
-                    </StrategyExecutionItem>
-                );
-            }
-
-            case 'string': {
-                const value = parseParameterString(parameters[name]);
-
-                return (
-                    <StrategyExecutionItem
-                        key={key}
-                        type={typeItem}
-                        values={value === '' ? undefined : [value]}
-                    >
-                        {value === '' ? 'is an empty string' : 'is set to'}
-                    </StrategyExecutionItem>
-                );
-            }
-
-            case 'number': {
-                const value = parseParameterNumber(parameters[name]);
-                return (
-                    <StrategyExecutionItem
-                        key={key}
-                        type={typeItem}
-                        values={[`${value}`]}
-                    >
-                        is a number set to
-                    </StrategyExecutionItem>
-                );
-            }
-
-            case 'default':
-                return null;
-        }
-
-        return null;
-    };
-
-    return useMemo(
-        () => ({
-            isCustomStrategy,
-            customStrategyItems: definition?.parameters
-                ?.map(mapCustomStrategies)
-                .filter(Boolean),
-        }),
-        [definition, isCustomStrategy, parameters],
-    );
-};
-
 type StrategyExecutionProps = {
     strategy: IFeatureStrategyPayload | CreateFeatureStrategySchema;
     displayGroupId?: boolean;
@@ -270,10 +70,8 @@ const NewStrategyExecution: FC<StrategyExecutionProps> = ({
     displayGroupId = false,
 }) => {
     const { strategies } = useStrategies();
-    const { isCustomStrategy, customStrategyItems } = useCustomStrategyItems(
-        strategy,
-        strategies,
-    );
+    const { isCustomStrategy, customStrategyParameters: customStrategyItems } =
+        useCustomStrategyParameters(strategy, strategies);
     const strategyParameters = useStrategyParameters(strategy, displayGroupId);
     const { constraints } = strategy;
